+++ src/main/java/com/mjhylkema/TeleportMaps/components/adventureLog/MinecartMap.java
@@ -1,7 +1,8 @@
-package com.mjhylkema.TeleportMaps.components;
+package com.mjhylkema.TeleportMaps.components.adventureLog;
 
 import com.mjhylkema.TeleportMaps.TeleportMapsConfig;
 import com.mjhylkema.TeleportMaps.TeleportMapsPlugin;
+import com.mjhylkema.TeleportMaps.components.BaseMap;
 import com.mjhylkema.TeleportMaps.definition.MinecartDefinition;
 import com.mjhylkema.TeleportMaps.ui.AdventureLogEntry;
 import com.mjhylkema.TeleportMaps.ui.UIHotkey;
